# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "SkillTree API",
		"description": "API for managing skill trees, skill nodes, and user progress tracking",
		"version": "1.0.0",
		"contact": [
			{
				"name": "Aditya Tripathi",
				"email": "atri0048@student.monash.edu"
			}
		]
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development server"
		}
	],
	"security": [
		{
			"bearerAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"ApiResponse": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"message": {
						"oneOf": [
							{
								"type": "string"
							},
							{
								"type": "object"
							}
						]
					},
					"status": {
						"type": "integer"
					}
				}
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"xpPoint": {
						"type": "integer"
					}
				}
			},
			"Tag": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"isRestricted": {
						"type": "boolean"
					},
					"restrictionDescription": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"SkillNode": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"xpPoint": {
						"type": "integer"
					},
					"parentNode": {
						"$ref": "#/components/schemas/SkillNodeBasic"
					},
					"childNode": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SkillNodeBasic"
						}
					},
					"skillNodeUser": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SkillNodeUser"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"SkillNodeBasic": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					}
				}
			},
			"SkillNodeUser": {
				"type": "object",
				"properties": {
					"xpPoint": {
						"type": "integer"
					}
				}
			},
			"SkillTree": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"creator": {
						"$ref": "#/components/schemas/User"
					},
					"tags": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Tag"
						}
					},
					"skillNodes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SkillNode"
						}
					},
					"skillTreeUser": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SkillTreeUser"
						}
					},
					"_count": {
						"type": "object",
						"properties": {
							"skillNodes": {
								"type": "integer"
							},
							"skillTreeUser": {
								"type": "integer"
							}
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"SkillTreeUser": {
				"type": "object",
				"properties": {
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"role": {
						"type": "string",
						"enum": ["ADMIN", "MEMBER"]
					},
					"verificationStatus": {
						"type": "string",
						"enum": ["PENDING", "VERIFIED"]
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				}
			},
			"CreateSkillTreeDto": {
				"type": "object",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1
					},
					"description": {
						"type": "string"
					},
					"tagIds": {
						"type": "array",
						"items": {
							"type": "string",
							"format": "uuid"
						}
					}
				}
			},
			"UpdateSkillTreeDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1
					},
					"description": {
						"type": "string"
					},
					"tagIds": {
						"type": "array",
						"items": {
							"type": "string",
							"format": "uuid"
						}
					}
				}
			},
			"UpdateSkillTreeUserDto": {
				"type": "object",
				"properties": {
					"role": {
						"type": "string",
						"enum": ["ADMIN", "MEMBER"]
					},
					"verificationStatus": {
						"type": "string",
						"enum": ["PENDING", "VERIFIED"]
					}
				}
			},
			"Error": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "integer"
					}
				}
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Unauthorized - Invalid or missing JWT token",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"ok": false,
							"message": "Unauthorized",
							"status": 401
						}
					}
				}
			},
			"Forbidden": {
				"description": "Forbidden - Insufficient permissions",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"ok": false,
							"message": "You must be an admin to perform this action",
							"status": 403
						}
					}
				}
			},
			"NotFound": {
				"description": "Resource not found",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"ok": false,
							"message": "Skill tree not found",
							"status": 404
						}
					}
				}
			},
			"Conflict": {
				"description": "Conflict - Resource already exists",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"ok": false,
							"message": "You are already a member of this skill tree",
							"status": 409
						}
					}
				}
			},
			"InternalServerError": {
				"description": "Internal server error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"ok": false,
							"message": "Internal server error",
							"status": 500
						}
					}
				}
			}
		}
	},
	"paths": {
		"/skilltree": {
			"get": {
				"tags": ["SkillTree - Public"],
				"summary": "Get all skill trees (public)",
				"description": "Fetch all public skill trees. Excludes restricted skill trees.",
				"security": [],
				"responses": {
					"200": {
						"description": "List of skill trees",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SkillTree"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"post": {
				"tags": ["SkillTree - Authenticated"],
				"summary": "Create a new skill tree",
				"description": "Create a new skill tree. Creator automatically becomes admin.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSkillTreeDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Skill tree created successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTree"
												}
											}
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid input data"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/{id}": {
			"get": {
				"tags": ["SkillTree - Public"],
				"summary": "Get skill tree by ID (public)",
				"description": "Fetch a specific skill tree by ID. Limited access to restricted content.",
				"security": [],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"200": {
						"description": "Skill tree details",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTree"
												}
											}
										}
									]
								}
							}
						}
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"patch": {
				"tags": ["SkillTree - Authenticated"],
				"summary": "Update skill tree",
				"description": "Update an existing skill tree. Must be admin of the skill tree.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSkillTreeDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Skill tree updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTree"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"delete": {
				"tags": ["SkillTree - Authenticated"],
				"summary": "Delete skill tree",
				"description": "Delete a skill tree and all related data. Must be admin of the skill tree.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"204": {
						"description": "Skill tree deleted successfully"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/auth/all": {
			"get": {
				"tags": ["SkillTree - Authenticated"],
				"summary": "Get all skill trees (authenticated)",
				"description": "Fetch all skill trees with user-specific information including restricted access.",
				"responses": {
					"200": {
						"description": "List of skill trees with user context",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SkillTree"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/auth/{id}": {
			"get": {
				"tags": ["SkillTree - Authenticated"],
				"summary": "Get skill tree by ID (authenticated)",
				"description": "Fetch detailed skill tree information with user context and progress.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"200": {
						"description": "Detailed skill tree with user context",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTree"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/{id}/join": {
			"post": {
				"tags": ["SkillTree - Membership"],
				"summary": "Join a skill tree",
				"description": "Join a skill tree as a member. Status will be VERIFIED for public trees, PENDING for restricted trees.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully joined the skill tree",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTreeUser"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"409": {
						"$ref": "#/components/responses/Conflict"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/{id}/leave": {
			"delete": {
				"tags": ["SkillTree - Membership"],
				"summary": "Leave a skill tree",
				"description": "Leave a skill tree. Creator cannot leave their own skill tree.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"204": {
						"description": "Successfully left the skill tree"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/{id}/members": {
			"get": {
				"tags": ["SkillTree - Membership"],
				"summary": "Get skill tree members",
				"description": "Get all members of a skill tree. Must be a member to view.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					}
				],
				"responses": {
					"200": {
						"description": "List of skill tree members",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SkillTreeUser"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/{id}/members/{userId}": {
			"patch": {
				"tags": ["SkillTree - Membership"],
				"summary": "Update user role or verification status",
				"description": "Update a user's role or verification status. Must be admin of the skill tree.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "User ID"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSkillTreeUserDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User membership updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"$ref": "#/components/schemas/SkillTreeUser"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"delete": {
				"tags": ["SkillTree - Membership"],
				"summary": "Remove user from skill tree",
				"description": "Remove a user from the skill tree. Must be admin. Cannot remove creator.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Skill tree ID"
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "User ID"
					}
				],
				"responses": {
					"204": {
						"description": "User removed from skill tree successfully"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/user/my-skilltrees": {
			"get": {
				"tags": ["SkillTree - User"],
				"summary": "Get my skill trees",
				"description": "Get all skill trees the authenticated user is a member of.",
				"responses": {
					"200": {
						"description": "List of user's skill trees",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SkillTreeUser"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/skilltree/user/{userId}/skilltrees": {
			"get": {
				"tags": ["SkillTree - User"],
				"summary": "Get user's skill trees",
				"description": "Get all skill trees a specific user is a member of.",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "User ID"
					}
				],
				"responses": {
					"200": {
						"description": "List of user's skill trees",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"properties": {
												"message": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/SkillTreeUser"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		}
	},
	"tags": [
		{
			"name": "SkillTree - Public",
			"description": "Public endpoints that don't require authentication"
		},
		{
			"name": "SkillTree - Authenticated",
			"description": "Authenticated endpoints for skill tree CRUD operations"
		},
		{
			"name": "SkillTree - Membership",
			"description": "Endpoints for managing skill tree memberships"
		},
		{
			"name": "SkillTree - User",
			"description": "User-specific skill tree endpoints"
		}
	]
}